requirements.txt
appdirs==1.4.3
asgiref==3.7.2
async-timeout==4.0.3
awscli==1.18.6
backports.zoneinfo==0.2.1
botocore==1.15.6
colorama==0.4.3
distlib==0.3.0
Django==4.2.8
djangorestframework==3.14.0
docutils==0.15.2
filelock==3.0.12
jmespath==0.9.5
psycopg2-binary==2.9.9
pyasn1==0.4.8
python-dateutil==2.8.1
pytz==2023.3.post1
PyYAML==5.2
redis==5.0.1
rsa==3.4.2
s3transfer==0.3.3
six==1.14.0
sortedcontainers==2.4.0
sqlparse==0.4.4
typing-extensions==4.9.0
urllib3==1.25.8
uWSGI==2.0.23
virtualenv==20.0.4
